Add exit bool to prevent improper reconnections (#2)
This commit fixes a bug in the library wherein all types of disconnects will
trigger a forced reconnection. This is not a behavior that we want since
there will be times where we trigger a disconnect and stop completely.

The DISCONNECT packet from MQTT is shared between both the client and broker.
i.e. The client can send a DISCONNECT packet to terminate connection to the MQTT broker, and the
broker can send a DISCONNECT packet to terminate the client's connection to the MQTT broker.
The newly added sync bool should make it easier to distinguish between a proper disconnection
vs a net or broker disconnection.

* Add an `exit` bool to track client disconnection requests